# enable the SLP Vectorizer optimization pass by default