@KristofferC 2018-03-23
@nanosoldier `runbenchmarks(ALL, vs = ":master")`