@KristofferC 2018-03-23
Meant to run this:

@nanosoldier `runbenchmarks(ALL, vs = "@f0087141b79736beec7b5a2ee946d5c4ec257167")`