@KristofferC 2018-03-23
The test run is supposed to show compilation time.