@ararslan 2018-03-24
Same error, but it looks like other requests have gotten through, so I'll just try again. GitHub.jl needs to be updated to work with the changes in HTTP.jl, which is why the error keeps happening.

@nanosoldier `runbenchmarks(ALL, vs="@f0087141b79736beec7b5a2ee946d5c4ec257167")`