@KristofferC 2018-03-25
Perhaps needs a restart @ararslan?