@andyferris 2018-03-26
I definitely think we should do this, though I've been confused over time as to what happens when this is off (e.g. when I don't use `-O3`) since I can still see `xmm` and `ymm` registers and so-on being used - are there other parts of LLVM or the Julia compiler that would likely make that happen?