@ararslan 2018-03-26
I modified the logging level in the server, so if this fails, hopefully we'll have a better understanding of why. Sorry for the noise here, Kristoffer, and thanks for your patience.

@nanosoldier `runbenchmarks(ALL, vs="@f0087141b79736beec7b5a2ee946d5c4ec257167")`