@ararslan 2018-03-26
I had to modify the server's local clone of HTTP to fix an `UndefVarError`. I'll submit a PR for that to HTTP, but in the meantime:

@nanosoldier `runbenchmarks(ALL, vs="@f0087141b79736beec7b5a2ee946d5c4ec257167")`