@ararslan 2018-03-27
This PR hit the same error again, and unfortunately Nanosoldier is now down until further notice for on-site work.