@KristofferC 2018-03-29
@nanosoldier `runbenchmarks(ALL, vs="@c12922eeea1afb59d05477698d408e4ff54ff7f1")`