@nanosoldier 2018-03-30
[Your benchmark job](https://github.com/JuliaLang/julia/pull/26594#issuecomment-377388906) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/ca881ebb8612578afe21f1390b0f66d01fd0681f/d295c82_vs_c12922e/report.md). cc @ararslan