@KristofferC 2018-03-30
Maybe I should rebase this? I don't think it is running om the merge commit.