@ararslan 2018-03-30
It should always be running on the merge commit unless there are branch conflicts.