@KristofferC 2018-03-30
Okay but I m quite sure it doesn't because e.g.the memory regressions from https://github.com/JuliaLang/julia/pull/26435#issuecomment-374804516 also shows up here which mean that this commit includes that PR but not the one that we compared against.