@KristofferC 2018-03-30
@nanosoldier `runbenchmarks(ALL, vs = ":vc/llvm6")`

let's try this