@ararslan 2018-03-30
Nanosoldier seems to be consistently hitting the `IOError` again.