@KristofferC 2018-03-30
https://github.com/JuliaWeb/GitHub.jl/pull/108 maybe :/