@nanosoldier 2018-03-31
[Your benchmark job](https://github.com/JuliaLang/julia/pull/26594#issuecomment-377644692) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/c37dcc3c798778ab5c473e01edddeadfe34ecddb/2701ae7_vs_91baf58/report.md). cc @ararslan