@KristofferC 2018-04-18
@nanosoldier `runbenchmarks(ALL, vs = ":master")`