@vchuravy 2018-05-06
rebase onto master, and then LGTM