@KristofferC 2018-05-06
@nanosoldier `runbenchmarks(ALL, vs = ":master")`