@nanosoldier 2018-05-06
[Your benchmark job](https://github.com/JuliaLang/julia/pull/26594#issuecomment-386863520) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/a4090c4cd8a514aab164f8c56a155882a0582486/93f7a16_vs_bfb1c1b/report.md). cc @ararslan