@vchuravy 2018-05-06
CI failures are all network failures.